+++ app/[locale]/oauth/[uid]/OAuthConsent.tsx
